GUI: Move QRImageWidget to its own file-pair
